+++ commands/images/achievement.js
@@ -1,12 +1,12 @@
-const Discord = require('discord.js');
-const AlexAPI = require('alexflipnote.js');
-const settings = require('../../settings.json');
-const AlexClient = new AlexAPI(process.env.apitoken);
+const Discord = require('discord.js'),
+  AlexAPI = require('alexflipnote.js'),
+  settings = require('../../settings.json'),
+  AlexClient = new AlexAPI(process.env.apitoken);
 
 exports.run = async (client, message, args) => {
-  let avatar = message.content.split(" ").slice(1, 2)
-  let text = message.content.split(" ").slice(2).join(" ")
-  let link = await AlexClient.image.achievement({text: text})
+  let avatar = message.content.split(" ").slice(1, 2),
+    text = message.content.split(" ").slice(2).join(" "),
+    link = await AlexClient.image.achievement({text: text})
 await message.channel.send({files: [{ attachment: link }]});
 };