+++ commands/images/amiajoke.js
@@ -1,11 +1,11 @@
-const Discord = require('discord.js');
-const AlexAPI = require('alexflipnote.js');
-const settings = require('../../settings.json');
-const AlexClient = new AlexAPI(process.env.apitoken);
+const Discord = require('discord.js'),
+  AlexAPI = require('alexflipnote.js'),
+  settings = require('../../settings.json'),
+  AlexClient = new AlexAPI(process.env.apitoken);
 
 exports.run = async (client, message, args) => {
-  let avatar = message.mentions.users.size ? message.mentions.users.first().avatarURL({ format: 'png', dynamic: true, size: 2048 }) : message.author.avatarURL({ format: 'png', dynamic: true, size: 2048 });
-  let link = await AlexClient.image.amiajoke({image: avatar})
+  let avatar = message.mentions.users.size ? message.mentions.users.first().avatarURL({ format: 'png', dynamic: true, size: 2048 }) : message.author.avatarURL({ format: 'png', dynamic: true, size: 2048 }),
+    link = await AlexClient.image.amiajoke({image: avatar})
 message.channel.send({files: [{ attachment: link }]});
 };