+++ commands/info/bug.js
@@ -1,10 +1,10 @@
-const Discord = require('discord.js');
-const superagent = require('superagent');
+const Discord = require('discord.js'),
+  superagent = require('superagent');
 
 exports.run = async (client, message, args, tools) => {
     if (!args[0]) return message.reply('Bruh, please give me a real bug, not thin air!');;
     if (args[0] === "suggestion") return message.reply("BUG! Not a suggestion...");
-    content = args.join(" ");
+    let content = args.join(" ");
     message.reply("Thanks for reporting a bug! Please don\'t spam this command.");
     let suggestEmbed = new Discord.MessageEmbed()
     .setColor("#d000a8")