+++ events/message.js
@@ -1,48 +1,43 @@
-const Discord = require('discord.js');
-const settings = require('../settings.json');
-const fs = require("fs");
-const superagent = require('superagent');
-const db = require('../database.json');
-const mongoose = require('mongoose');
-const Settings = require("../models/settings.js")
-const humanizeDuration = require("humanize-duration");
-
-const cooldowns = new Discord.Collection();
+const Discord = require('discord.js'),
+  settings = require('../settings.json'),
+  fs = require("fs"),
+  superagent = require('superagent'),
+  mongoose = require('mongoose'),
+  Settings = require("../models/settings.js"),
+  humanizeDuration = require("humanize-duration"),
+  cooldowns = new Discord.Collection();
 
 module.exports = async message => {
-
 
     let guildSettings = await Settings.findOne({
         guildID: message.guild.id
     })
     if (message.channel.type === "dm") return;
-    if (message.author.bot) return;
+    if (message.author.bot || message.webhookID) return;
 
 
     let client = message.client;
     if (!guildSettings) {
         let prefix = settings.prefix;
         if (!message.content.startsWith(prefix)) return;
-        let command = message.content.split(' ')[0].slice(prefix.length);
-        let params = message.content.split(' ').slice(1);
-        let perms = client.elevation(message);
-        let cmd;
+        let command = message.content.split(' ')[0].slice(prefix.length),
+          params = message.content.split(' ').slice(1),
+          perms = client.elevation(message),
+          cmd;
 
 
-        if (client.commands.has(command)) {
+        if (client.commands.has(command))
             cmd = client.commands.get(command);
-        } else if (client.aliases.has(command)) {
+        else if (client.aliases.has(command))
             cmd = client.commands.get(client.aliases.get(command));
-        }
         if (cmd) {
-            if (!cooldowns.has(cmd.help.name)) {
+            if (!cooldowns.has(cmd.help.name))
                 cooldowns.set(cmd.help.name, new Discord.Collection());
-            }
 
-            const now = Date.now();
-            const timestamps = cooldowns.get(cmd.help.name);
-            const cooldownAmount = cmd.conf.cooldown * 1000 || 0;
-if (message.author.id != settings.ownerid) {
+            const now = Date.now(),
+              timestamps = cooldowns.get(cmd.help.name),
+              cooldownAmount = cmd.conf.cooldown * 1000 || 0;
+            if (message.author.id != settings.ownerid) {
             if (timestamps.has(message.author.id)) {
                 const expirationTime = timestamps.get(message.author.id) + cooldownAmount;
 
@@ -55,8 +50,7 @@ if (message.author.id != settings.ownerid) {
             timestamps.set(message.author.id, now);
             setTimeout(() => timestamps.delete(message.author.id), cooldownAmount);
             if (perms < cmd.conf.permLevel) {
-                console.log(`Command: ${settings.prefix}` + cmd.help.name + " - Guild: " + message.guild.name + " ID: " + message.guild.id)
-                return;
+                return console.log(`Command: ${settings.prefix + cmd.help.name} - Guild: ${message.guild.name} ID: ${message.guild.id}`)
             }
             try {
                 cmd.run(client, message, params, perms);
@@ -69,30 +63,28 @@ if (message.author.id != settings.ownerid) {
                 message.react('811296689783832617')
             }
         }
-        else { message.react('811294703626223687') }
+        else message.react('811294703626223687');
     } else {
         let prefix1 = guildSettings.prefix;
         if (!message.content.startsWith(prefix1)) return;
-        let command = message.content.split(' ')[0].slice(prefix1.length);
-        let params = message.content.split(' ').slice(1);
-        let perms = client.elevation(message);
-        let cmd;
+        let command = message.content.split(' ')[0].slice(prefix1.length),
+          params = message.content.split(' ').slice(1),
+          perms = client.elevation(message),
+          cmd;
 
 
-        if (client.commands.has(command)) {
+        if (client.commands.has(command))
             cmd = client.commands.get(command);
-        } else if (client.aliases.has(command)) {
+        else if (client.aliases.has(command))
             cmd = client.commands.get(client.aliases.get(command));
-        }
         if (cmd) {
-            if (!cooldowns.has(cmd.help.name)) {
+            if (!cooldowns.has(cmd.help.name))
                 cooldowns.set(cmd.help.name, new Discord.Collection());
-            }
 
-            const now = Date.now();
-            const timestamps = cooldowns.get(cmd.help.name);
-            const cooldownAmount = cmd.conf.cooldown * 1000 || 0;
-if (message.author.id != settings.ownerid) {
+            const now = Date.now(),
+              timestamps = cooldowns.get(cmd.help.name),
+              cooldownAmount = cmd.conf.cooldown * 1000 || 0;
+            if (message.author.id != settings.ownerid) {
 
             if (timestamps.has(message.author.id)) {
                 const expirationTime = timestamps.get(message.author.id) + cooldownAmount;
@@ -104,8 +96,7 @@ if (message.author.id != settings.ownerid) {
             }}
 
             if (perms < cmd.conf.permLevel) {
-                console.log(`Command: ${guildSettings.prefix}` + cmd.help.name + " - Guild: " + message.guild.name + " ID: " + message.guild.id)
-                return;
+                return console.log(`Command: ${guildSettings.prefix}` + cmd.help.name + " - Guild: " + message.guild.name + " ID: " + message.guild.id)
 
             }
             try {
@@ -121,9 +112,7 @@ if (message.author.id != settings.ownerid) {
                 console.log(err)
                 message.react('811296689783832617')
             }
-        } else { message.react('811294703626223687') }
+        } else message.react('811294703626223687');
 
     }
-
-
 };