+++ events/ready.js
@@ -1,6 +1,4 @@
 const chalk = require('chalk');
-const Discord = require("discord.js");
-const fs = require('fs');
 module.exports = client => {
   client.user.setActivity(`${client.guilds.cache.size} servers  ⋙ c!help`, { type: 'WATCHING' })
   setInterval(() => {