+++ server.js
@@ -1,5 +1,5 @@
-const express = require('express');
-const server = express();
+const express = require('express'),
+  server = express();
 server.all('/', (req, res)=>{
     res.send('CommandStorm Discord Bot! ')
 })