+++ util/eventLoader.js
@@ -1,4 +1,4 @@
-const reqEvent = (event) => require(`../events/${event}`);
+const reqEvent = event => require(`../events/${event}`);
 module.exports = client => {
   client.on('ready', () => reqEvent('ready')(client));
   client.on('reconnecting', () => reqEvent('reconnecting')(client));